Move components to more appropriate spot
