+++ osu.Game.Rulesets.Osu/Edit/OsuChecker.cs
@@ -5,8 +5,8 @@ using System.Collections.Generic;
 using System.Linq;
 using osu.Game.Beatmaps;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Edit.Checks.Components;
 using osu.Game.Rulesets.Osu.Edit.Checks;
-using osu.Game.Screens.Edit.Verify.Components;
 
 namespace osu.Game.Rulesets.Osu.Edit
 {
