+++ osu.Game/Rulesets/Edit/Checks/Components/CheckMetadata.cs
@@ -1,7 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Screens.Edit.Verify
+namespace osu.Game.Rulesets.Edit.Checks.Components
 {
     public class CheckMetadata
     {
